add validation to add one&range appointments

+++ DocHub.Core/DTO/AppointmentAddRequest.cs
@@ -1,16 +1,19 @@
 using System.ComponentModel.DataAnnotations;
 using DocHub.Core.Domain.Entities;
+using DocHub.Core.Domain.RepositoryContracts;
+using DocHub.Core.ValidationAttributes;
 
 namespace DocHub.Core.DTO;
 
 public class AppointmentAddRequest
 {
-    [Required]
-    public DateTime Start { get; set; }
+    [Required] public DateTime Start { get; set; }
+
     [RegularExpression(@"^\d+$", ErrorMessage = "Duration time must be numeric")]
     [Required(ErrorMessage = "The duration of the visit is required")]
     [Range(1, 1400, ErrorMessage = "Duration time must be greater than 0")]
     public required int Duration { get; set; }
+
     [RegularExpression(@"^([01]?[0-9]|2[0-3]):[0-5][0-9]$", ErrorMessage = "The visit time must be a clock value")]
     public required string Hour { get; set; }
 
@@ -23,8 +26,7 @@ public class AppointmentAddRequest
         return new Appointment()
         {
             Start = startDate,
-            End = startDate.AddMinutes((double) Duration)
+            End = startDate.AddMinutes(Duration)
         };
     }
-    
 }
